功能优化和统计部分完成

+++ src/main/java/com/itheima/service/impl/DeptServiceImpl.java
@@ -1,10 +1,13 @@
 package com.itheima.service.impl;
 
+import com.itheima.exception.BusinessException;
 import com.itheima.mapper.DeptMapper;
+import com.itheima.mapper.EmpMapper;
 import com.itheima.pojo.Dept;
 import com.itheima.service.DeptService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -13,12 +16,24 @@ import java.util.List;
 public class DeptServiceImpl implements DeptService {
     @Autowired
     private DeptMapper deptMapper;
+    @Autowired
+    private EmpMapper empMapper;
+
+    @Override
     public List<Dept> findAll(){
         return deptMapper.findAll();
     }
+
+    @Transactional(rollbackFor = Exception.class)
+    @Override
     public void deleteById(Integer id) {
+        if(empMapper.countEmpByDeptId(id) > 0){
+            throw new BusinessException("对不起，当前部门下有员工，不能直接删除！");
+        }
         deptMapper.deleteById(id);
     }
+    @Transactional(rollbackFor = Exception.class)
+    @Override
     public void save(Dept dept) {
         //补全基础属性
         dept.setCreateTime(LocalDateTime.now());
@@ -26,9 +41,14 @@ public class DeptServiceImpl implements DeptService {
         //保存部门
         deptMapper.insert(dept);
     }
+
+    @Override
     public Dept getById(Integer id) {
         return deptMapper.getById(id);
     }
+
+    @Transactional(rollbackFor = Exception.class)
+    @Override
     public void update(Dept dept) {
         //补全基础属性
         dept.setUpdateTime(LocalDateTime.now());
